Alterações com Jupter

+++ src/test/java/br/com/proway/senior/cargosESalarios/model/GrauInstrucaoDaoTest.java
@@ -1,13 +1,15 @@
 package br.com.proway.senior.cargosESalarios.model;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 import java.util.ArrayList;
 
 import org.junit.Before;
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 class GrauInstrucaoDaoTest {
 
@@ -23,7 +25,7 @@ class GrauInstrucaoDaoTest {
 	void testCreate() {
 		dao.limparArray();
 		GrauInstrucaoModel gi = new GrauInstrucaoModel("Ensino Médio");
-		assertEquals(0, dao.create(gi));
+		assertEquals((Integer)0, dao.create(gi));
 	}
 	
 	@Test
